+++ cmd/licenses.go
@@ -19,14 +19,17 @@ func init() {
 func runLicensesCmd(_ *cobra.Command, _ []string) {
 	fmt.Println(`
 The source code for slackbridge is available to you under the terms of the MIT
-License, and can be found at https://github.com/ahamlinman/slackbridge.
+License, and can be found at the following location:
+
+  https://github.com/ahamlinman/slackbridge
 
 This release of slackbridge also incorporates code from the go-multierror and
 errwrap libraries distributed by HashiCorp, Inc. under the terms of the Mozilla
-Public License, version 2.0. The source code of these libraries is available to
-you at the following locations:
+Public License, version 2.0. This source code is available to you at the above
+location under the "vendor/github.com/hashicorp/" directory, or at the
+following locations:
 
-  - go-multierror: https://github.com/hashicorp/go-multierror
-  - errwrap: https://github.com/hashicorp/errwrap
+  go-multierror: https://github.com/hashicorp/go-multierror
+  errwrap: https://github.com/hashicorp/errwrap
 `)
 }