+++ CHANGELOG.md
@@ -13,6 +13,8 @@ project adheres to a scheme based on [Semantic Versioning] as follows:
 [Semantic Versioning]: http://semver.org/spec/v2.0.0.html
 
 ## [Unreleased]
+
+## [v0.1.3] - 2018-02-24
 ### Changed
 - Package slackio has moved to https://github.com/ahamlinman/slackio
 - All dependencies are now version-controlled to support better reproducibility
@@ -42,7 +44,8 @@ project adheres to a scheme based on [Semantic Versioning] as follows:
 - Initial versioned release
 - Changelog for all notable modifications going forward
 
-[Unreleased]: https://github.com/ahamlinman/slackbridge/compare/v0.1.2...HEAD
+[Unreleased]: https://github.com/ahamlinman/slackbridge/compare/v0.1.3...HEAD
+[v0.1.3]: https://github.com/ahamlinman/slackbridge/tree/v0.1.3
 [v0.1.2]: https://github.com/ahamlinman/slackbridge/tree/v0.1.2
 [v0.1.1]: https://github.com/ahamlinman/slackbridge/tree/v0.1.1
 [v0.1.0]: https://github.com/ahamlinman/slackbridge/tree/v0.1.0